refactor: 컨벤에 맞게 변경
