+++ src/components/common/progress-with-label/progress-with-label.tsx
@@ -1,9 +1,17 @@
 import { cva } from 'class-variance-authority';
-import { User, Users } from 'lucide-react';
+import { Users } from 'lucide-react';
 
 import { Field, FieldLabel } from '@/components/ui/field/field';
-import type { ProgressProps } from '@/components/ui/progress-bar/index';
-import { Progress } from '@/components/ui/progress-bar/index';
+import { Progress, type ProgressProps } from '@/components/ui/progress-bar';
+
+const progressStateClasses = cva('ml-auto font-bold', {
+  variants: {
+    variant: {
+      groupEat: 'text-sosoeat-orange-500',
+      groupBuy: 'text-sosoeat-blue-500',
+    },
+  },
+});
 
 interface ProgressWithLabelProps {
   current: number;
@@ -11,24 +19,17 @@ interface ProgressWithLabelProps {
   variant: ProgressProps['variant'];
 }
 
-export function ProgressWithLabel({ current, max, variant }: ProgressWithLabelProps) {
+export const ProgressWithLabel = ({ current, max, variant }: ProgressWithLabelProps) => {
   const clampedCurrent = Math.min(Math.max(0, current), max);
   const progress = (clampedCurrent / max) * 100;
-
-  const progressStateClasses = cva('ml-auto font-bold', {
-    variants: {
-      variant: {
-        groupEat: 'text-sosoeat-orange-500',
-        groupBuy: 'text-sosoeat-blue-500',
-      },
-    },
-  });
 
   return (
     <Field className="w-full max-w-sm">
       <FieldLabel htmlFor="progress-upload">
-        <span className="m-0 flex items-end justify-center gap-1">{<Users />}</span>
-        <span className="text-bold -ml-1">{`${clampedCurrent}/${max} 참여중`}</span>
+        <span className="m-0 flex items-end justify-center gap-1">
+          <Users />
+        </span>
+        <span className="-ml-1 font-bold">{`${clampedCurrent}/${max} 참여중`}</span>
         <span className={progressStateClasses({ variant })}>
           {clampedCurrent === max ? '마감' : ``}
         </span>
@@ -36,4 +37,4 @@ export function ProgressWithLabel({ current, max, variant }: ProgressWithLabelPr
       <Progress value={progress} variant={variant} id="progress-upload" />
     </Field>
   );
-}
+};
