+++ src/components/ui/progress-bar/progress.tsx
@@ -1,11 +1,10 @@
 'use client';
 
-import * as React from 'react';
-
 import { Progress as ProgressPrimitive } from 'radix-ui';
 
-import { ProgressProps } from '@/components/ui/progress-bar/progress.type';
 import { cn } from '@/lib/utils';
+
+import type { ProgressProps } from './progress.types';
 
 const variantStyles = {
   groupBuy: 'bg-sosoeat-blue-500',
@@ -18,14 +17,14 @@ const variantStylesFull = {
   error: 'bg-red-600',
 };
 
-function Progress({ className, value, variant, ...props }: ProgressProps) {
+export const Progress = ({ className, value, variant, ...props }: ProgressProps) => {
   const isFull = (value ?? 0) >= 100;
   return (
     <ProgressPrimitive.Root
       data-slot="progress"
       value={value}
       className={cn(
-        'bg-muted relative flex h-2 w-[328px] items-center overflow-x-hidden rounded-full',
+        'bg-muted relative flex h-2 w-82 items-center overflow-x-hidden rounded-full',
         className
       )}
       {...props}
@@ -40,6 +39,4 @@ function Progress({ className, value, variant, ...props }: ProgressProps) {
       />
     </ProgressPrimitive.Root>
   );
-}
-
-export { Progress };
+};
