--- src/components/ui/progress-bar/progress.type.ts
@@ -1,5 +1,0 @@
-import { Progress as ProgressPrimitive } from 'radix-ui';
-
-export interface ProgressProps extends React.ComponentProps<typeof ProgressPrimitive.Root> {
-  variant: 'groupBuy' | 'groupEat';
-}
