+++ src/components/ui/progress-bar/progress.types.ts
@@ -1,0 +1,7 @@
+import type { ComponentProps } from 'react';
+
+import { Progress as ProgressPrimitive } from 'radix-ui';
+
+export interface ProgressProps extends ComponentProps<typeof ProgressPrimitive.Root> {
+  variant: 'groupBuy' | 'groupEat';
+}
